Update MoleId generation
Previously the Mole ID (corresponding to its X index and Y index)
started from 0 (so the first Mole would be 0000). This can be confusing
to people not used to code. This has been changed to start from 1 (so
now the first Mole ID would be 0101), resulting in an ID easier to
understand.