Add LaserCursor class
Originally, the cursor of the laser (the object placed at the tip of the
laser if it hit something) was a simple object controlled by the Pointer
class. It has now its own class to give more control over the cursor and
to allow custom cursors with their own behavior.