Fix invalid log datetime format
The datetime (time stamp) saved in the logs is generated by the
System.DateTime.Now() function. This function uses the local
configuration of the computer to generate a string. One particular
behavior is that it will adapt the string format depending of the
language of the computer, meaning that on certain computers, the
datetime format wasn't the planned DD/MM/YYYY hh:mm:ss format but had
the hours in the 12  hour AM/PM format. This unplanned format resulted
in the database refusing to save the logs.