Fix bad Mole reset and add disabled cooldown
Fixed multiple minor issues witl the Mole reset. Also added a "disabled
cooldown" to prevent a Mole to be re-enabled right after being disabled
(poped or expired). This could happen in some rare cases. Also fixed a
minor issue with the GameDirector and decreased the pointers cooldown
for smoother gameplay.