Add UI update when playing pattern
The modifiers part of the UI was lacking a functionality to update
itself automatically when a modifier was set (since the modifiers were
not initially designed to be modified from another source than the UI).
This could lead to conflicts. Now the UI updates itself when the
modifiers/speed are changed.