LaserMapper: Add Motor Space Calibration Support
Currently we need to setup the motor space manually.
It would be much more useful if the motor space could
be set up by the VR participant, depending on the
participants' abilities. This commit adds the first
initial implementation for motor space calibration. By
moving the controller around the participant can draw
how high and wide the motor space should be.