SampleLogger: Log The Visible Viewport
In order to truly understand what happened during
a game, we also need information about what was
visible during playthrough. The best way to do this
is to track the positions of some proxy objects at
the very boundary of the VR visual field. This is
not 100% accurate, but good enough for analysis for
the time being.