Reset position of the player at the start of the game
We decided to add a feature to reset the position, the height and the rotation of the player. This also places every player at the same height to facilitate data comparison.
The current problem that need more investigation is the Motor Space position, does the current position fit every person, and can it adapt to people siting with a table in front of them or in a wheelchair with armrests without disturbing their movements.