Refactoring of the Pattern System
Added some parameters in the Pattern system to make it more modular for future updates.
Theses new parameters makes it easier to add new paradigms and change how the game works and react to the player.
It's also easier to manipulate some Variables and States in the game wich can help to implement new features