OnBoarding and Mirror Therapy Bug Resolved
We had bugs which appeared during the OnBoarding 1 and Mirror Therapy
session. So we tried to resolve them and in order to avoid this issues
in the next time, we found the origin of this problem. For the Mirror
Therapy, the bug's origin is just that a function in a 'if' condition
wasn't indented, so when the condition was respected, it returned
nothing. For the OnBoarding Session, the coordinates of the next
distractor were the same as the mole, so the mole couldn't appear.
However, if a distractor become a mole, there are absolutly no problem.