BasicPointer: Added debug controls
Debug controls allow to play the game without a VR headset
by holding the Ctrl key and the arrow keys
or holding Ctrl+Alt keys and moving the mouse.
This mode only works if no VR headset is active.
The mouse speed is adjustable in Unity's inspector panel.
PointerControl() is the function that contains the VR controls.