Add EMG gesture manager and update VirtualHand prefab
Introduced EMGClassifiedGestureManager script and meta file. Updated VirtualHand prefab with new right hand skeleton structure and components.

EMGClassifiedGestureManager needs only the references to the Skeleton Poser of the hand model and the Scriptable Objects which the poses were saved as to work.

For testing purposes, Keyboard 1, 2, 3, and 4 trigger pose changes for now, to be deprecated once EMG classification becomes the main driver for state change.