VirtualHand now has proper proportions
Eliminated the distortion effect caused by the VirtualHand parent having a Z scale of 1.9, changed the visual aides to be a continuation of the gloved hand SteamVR model in VirtualHand.

EMGPointer.cs now has a Serialized Field variable called Hand Offset and SteamVRVisualHand, those are used to offset the hand (must be a negative number in current configuration) and to disable the SteamVR automatic hand model appearing to showcase a hand holding the tracker object, as well as the tracker object itself being automatically instantiated in the scene. This only happens when EMGPointer gets instantiated, and they get re-enabled on Disable();.